requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.9